+++ CHANGELOG.rst
@@ -4,7 +4,11 @@ CHANGELOG
 5.0.0a11 (unreleased)
 ---------------------
 
-- Nothing changed yet.
+- Cache JSONField schema validator object
+  [vangheem]
+
+- JSONField works with dict instead of requiring str(which is then converted to dict anyways)
+  [vangheem]
 
 
 5.0.0a10 (2019-06-27)