+++ CHANGELOG.rst
@@ -7,6 +7,12 @@ CHANGELOG
 - Documentation improvements
   [vangheem]
 
+- Cache JSONField schema validator object
+  [vangheem]
+
+- JSONField works with dict instead of requiring str(which is then converted to dict anyways)
+  [vangheem]
+
 
 5.0.0a10 (2019-06-27)
 ---------------------